allow crate library again
